test fixes, lint-driven tweaks
